tests: moved out fixture for canvas
